feat(ui): move scenes and debug to side menu
