+++ src/router/index.js
@@ -4,6 +4,8 @@ import ControlPanel from '@/components/ControlPanel'
 import Settings from '@/components/Settings'
 import Mesh from '@/components/Mesh'
 import Store from '@/components/Store'
+import Scenes from '@/components/Scenes'
+import Debug from '@/components/Debug'
 
 Vue.use(Router)
 
@@ -20,6 +22,18 @@ export default new Router({
       path: '/settings',
       name: 'Settings',
       component: Settings,
+      props: true
+    },
+    {
+      path: '/scenes',
+      name: 'Scenes',
+      component: Scenes,
+      props: true
+    },
+    {
+      path: '/debug',
+      name: 'Debug',
+      component: Debug,
       props: true
     },
     {
